Ftr: config api
