+++ config/tests/test_api.rs
@@ -20,7 +20,6 @@ mod tests_api {
     use std::env;
     use std::sync::Once;
 
-    use config::api::{dubbo_get_protocol, dubbo_set_protocol};
     use ctor::ctor;
 
     use config::get_root_config;
@@ -44,30 +43,9 @@ mod tests_api {
     }
 
     #[test]
-    fn test_api_overwrite_yaml() {
-        let root_config = get_root_config();
-        println!("{:#?}", root_config);
-        root_config
-            .lock()
-            .unwrap()
-            .protocols
-            .get_mut("dubbo")
-            .unwrap()
-            .port = "20987".to_string();
-        assert_eq!(
-            root_config
-                .lock()
-                .unwrap()
-                .protocols
-                .get("dubbo")
-                .unwrap()
-                .port,
-            "20987".to_string()
-        );
-    }
-
-    #[test]
     fn test_dubbo_set_protocol() {
+        let config_wrapper = get_root_config();
+
         dubbo_set_protocol("dubbo", "ip".to_string(), "122.22.22.22".to_string()).unwrap();
         dubbo_set_protocol("dubbo", "port".to_string(), "111".to_string()).unwrap();
         dubbo_set_protocol("dubbo", "name".to_string(), "dubbo".to_string()).unwrap();
